/taler/transfer reads NEXUS_PRODUCTION env.
If NEXUS_PRODUCTION is defined, the requests to /taler/transfer
goes through the ordinary EBICS layer; in contrast, when such env
variable is not defined, then Nexus mocks all the requests as
successful payments in its local database.  In this latter case,
no bank or sandbox are required for the nexus to participate in
tests.