More "history logic" to the Sandbox.
Upon receiving a pain.001 document, the sandbox
stores its details into the database and associates
the new record with the requesting subscriber.

As a consequence, the sandbox now queries the history
of payments related to the requesting subscriber.

Note: the resulting C53 response does NOT have all the
details from the payment YET.  The way it is, it
only tells _how many_ payments were initiated by the
requesting subscriber.